Refine explorer search shortcut
