+++ src/modules/explorer/ExplorerSearch.tsx
@@ -1,0 +1,195 @@
+import { Input } from "@/components/ui/input";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import {
+  Cancel01Icon,
+  Folder01Icon,
+  Search01Icon,
+} from "@hugeicons/core-free-icons";
+import { HugeiconsIcon } from "@hugeicons/react";
+import { invoke } from "@tauri-apps/api/core";
+import { motion } from "motion/react";
+import {
+  forwardRef,
+  useEffect,
+  useImperativeHandle,
+  useRef,
+  useState,
+} from "react";
+import { fileIconUrl } from "./lib/iconResolver";
+
+type SearchHit = {
+  path: string;
+  rel: string;
+  name: string;
+  is_dir: boolean;
+};
+
+type Props = {
+  rootPath: string;
+  onOpenFile: (path: string) => void;
+  open: boolean;
+  onActiveChange?: (active: boolean) => void;
+};
+
+export type ExplorerSearchHandle = {
+  focus: () => void;
+  isFocused: () => boolean;
+};
+
+export const ExplorerSearch = forwardRef<ExplorerSearchHandle, Props>(function ExplorerSearch({
+  rootPath,
+  onOpenFile,
+  open,
+  onActiveChange,
+}: Props,
+  ref,
+) {
+  const [query, setQuery] = useState("");
+  const [results, setResults] = useState<SearchHit[]>([]);
+  const [searching, setSearching] = useState(false);
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const active = query.trim().length > 0;
+
+  useEffect(() => {
+    onActiveChange?.(active);
+  }, [active, onActiveChange]);
+
+  useEffect(() => {
+    if (open) {
+      inputRef.current?.focus();
+    } else {
+      setQuery("");
+      setResults([]);
+      setSearching(false);
+    }
+  }, [open]);
+
+  useEffect(() => {
+    const q = query.trim();
+    if (!q) {
+      setResults([]);
+      setSearching(false);
+      return;
+    }
+    setSearching(true);
+    let alive = true;
+    const handle = setTimeout(async () => {
+      try {
+        const hits = await invoke<SearchHit[]>("fs_search", {
+          root: rootPath,
+          query: q,
+          limit: 200,
+        });
+        if (alive) setResults(hits);
+      } catch (e) {
+        if (alive) {
+          console.error("fs_search failed:", e);
+          setResults([]);
+        }
+      } finally {
+        if (alive) setSearching(false);
+      }
+    }, 120);
+
+    return () => {
+      alive = false;
+      clearTimeout(handle);
+    };
+  }, [query, rootPath]);
+
+  useImperativeHandle(
+    ref,
+    () => ({
+      focus: () => {
+        requestAnimationFrame(() => {
+          inputRef.current?.focus();
+        });
+      },
+      isFocused: () => document.activeElement === inputRef.current,
+    }),
+    [],
+  );
+
+  return (
+    <div className="flex flex-col">
+      {open ? (
+        <motion.div
+          className="relative shrink-0 px-2 py-1.5"
+          initial={{ opacity: 0, transform: "translateY(-15px)" }}
+          animate={{ opacity: 1, transform: "translateY(0px)" }}
+        >
+          <HugeiconsIcon
+            icon={Search01Icon}
+            size={13}
+            strokeWidth={2}
+            className="absolute top-1/2 left-4 -translate-y-1/2 text-muted-foreground"
+          />
+          <Input
+            ref={inputRef}
+            value={query}
+            onChange={(e) => setQuery(e.target.value)}
+            placeholder="Search files…"
+            className="h-7 pr-7 pl-6.5 text-xs"
+          />
+          {query ? (
+            <button
+              type="button"
+              onClick={() => setQuery("")}
+              className="absolute top-1/2 right-3.5 -translate-y-1/2 rounded p-0.5 text-muted-foreground hover:bg-accent hover:text-foreground"
+              aria-label="Clear search"
+            >
+              <HugeiconsIcon icon={Cancel01Icon} size={11} strokeWidth={2} />
+            </button>
+          ) : null}
+        </motion.div>
+      ) : null}
+
+      {active ? (
+        <ScrollArea className="min-h-0 flex-1">
+          <div className="py-1">
+            {searching && results.length === 0 ? (
+              <div className="px-3 py-2 text-[11px] text-muted-foreground">
+                Searching…
+              </div>
+            ) : results.length === 0 ? (
+              <div className="px-3 py-2 text-[11px] text-muted-foreground">
+                No matches
+              </div>
+            ) : (
+              results.map((hit) => {
+                const url = hit.is_dir ? null : fileIconUrl(hit.name);
+                return (
+                  <button
+                    key={hit.path}
+                    type="button"
+                    onClick={() => {
+                      if (!hit.is_dir) onOpenFile(hit.path);
+                    }}
+                    className="flex w-full items-center gap-1.5 px-2 py-1 text-left text-xs hover:bg-accent"
+                    title={hit.path}
+                  >
+                    {url ? (
+                      <img src={url} alt="" className="size-3.5 shrink-0" />
+                    ) : (
+                      <HugeiconsIcon
+                        icon={Folder01Icon}
+                        size={13}
+                        strokeWidth={1.75}
+                        className="shrink-0 text-muted-foreground"
+                      />
+                    )}
+                    <span className="truncate">{hit.name}</span>
+                    <span className="ml-auto truncate text-[10px] text-muted-foreground">
+                      {hit.rel}
+                    </span>
+                  </button>
+                );
+              })
+            )}
+          </div>
+        </ScrollArea>
+      ) : null}
+    </div>
+  );
+});
