+++ src/modules/explorer/index.ts
@@ -1,1 +1,2 @@
 export { FileExplorer } from "./FileExplorer";
+export { ExplorerSearch } from "./ExplorerSearch";
