+++ src/modules/shortcuts/shortcuts.ts
@@ -13,6 +13,7 @@ export type ShortcutId =
   | "tab.prev"
   | "tab.selectByIndex"
   | "search.focus"
+  | "explorer.search"
   | "ai.toggle"
   | "ai.askSelection"
   | "shortcuts.open"
@@ -90,6 +91,13 @@ export const SHORTCUTS: Shortcut[] = [
     label: "Jump to tab 1–9",
     group: "Tabs",
     defaultBindings: [{ [MOD_PROP]: true, key: "1" }],
+  },
+  {
+    id: "explorer.search",
+    label: "Search files",
+    keys: [MOD_KEY, SHIFT_KEY, "F"],
+    group: "Search",
+    match: (e) => isMod(e) && e.shiftKey && e.key.toLowerCase() === "f",
   },
   {
     id: "search.focus",
